Update deps
Update a few plugins, python stuff, and update my lazy lockfile

The new maintainer of none-ls has for whatever reason
decided in https://github.com/nvimtools/none-ls.nvim/issues/58 to
create a bunch of churn and deprecate some random linters or move them
to other repos. This is a good time to just drop null-ls and move to
conform and nvim-lint.