Support autoswitching to nightmode (#161)
Kitty at some point added support for automatically switching between
a day and night theme. I've added the right symlinks for that and redid
the themes to better match the original solarized colorschemes.

I also brought in an autodarkmode vim plugin to automagically set vim to
dark mode as well