Add script for commitmono fixing
I really like https://commitmono.com and wanted to try it out, but the
download process is broken if you want to customize the font (which
I did). I'm including both my custom json for configuring the site, and
a small fix script that modifies the otf font to make everything
recognize it as monospace.